requirements.txt
streamlit==1.32.0
pylatexenc==2.10
networkx==3.2.1
plotly==5.18.0
google-generativeai==0.4.0
python-docx==1.1.0
pypandoc==1.11
pdf2image==1.17.0
Pillow==10.2.0
streamlit-agraph==0.0.45
